requirements.txt
APScheduler==3.7.0
asgiref==3.4.1
bitcoin==1.1.39
blockcypher==1.0.92
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.6
django-countries==7.2.1
django-cryptocurrency-payment==0.1.1
django-phonenumber-field==5.2.0
forex-python==1.6
idna==3.2
merchant-wallet==0.1.1
phonenumbers==8.12.31
Pillow==8.3.1
python-dateutil==2.8.2
pytz==2021.1
requests==2.26.0
schedule==1.1.0
simplejson==3.17.5
six==1.16.0
sqlparse==0.4.1
tzlocal==2.1
urllib3==1.26.6
whitenoise==5.3.0
